RD-3949: Extracted form as an external component
